% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markin.R
\name{markin}
\alias{markin}
\alias{markin_inject}
\title{The Markdown Injector}
\usage{
markin(file, verbose = FALSE)

markin_inject(lines, barefile, verbose = FALSE)
}
\arguments{
\item{file}{The Markdown file}

\item{verbose}{If TRUE, verbose messages are outputted, otherwise not.}

\item{lines}{(character) The Markdown lines to be processed.}

\item{barefile}{(character string) The bare file without the filename extension.}
}
\value{
(invisible; character vector)
The Markdown file content with updated lines.

(character) The updated Markdown lines.
}
\description{
The Markdown Injector
}
