% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdi.R
\name{mdi_inject}
\alias{mdi_inject}
\title{The Markdown Injector (MDI)}
\usage{
mdi_inject(lines, barefile, path = ".", verbose = FALSE)
}
\arguments{
\item{verbose}{If TRUE, verbose messages are outputted, otherwise not.}

\item{file}{The Markdown file}
}
\value{
(invisible; character vector)
The Markdown file content with updated lines.
}
\description{
The Markdown Injector (MDI)
}
